Revert "Merge pull request #63632 from azat/global-query-profiler-followup"
This reverts commit 28ded4a471af3455696714553dd676396e332e98, reversing
changes made to 7bce45f2a02abe466b1f66c52772bc8527b48087.
