+++ src/Common/ThreadPool.h
@@ -242,11 +242,6 @@ public:
                 if (unlikely(global_profiler_real_time_period != 0 || global_profiler_cpu_time_period != 0))
                     thread_status.initGlobalProfiler(global_profiler_real_time_period, global_profiler_cpu_time_period);
             }
-            else
-            {
-                UNUSED(global_profiler_real_time_period);
-                UNUSED(global_profiler_cpu_time_period);
-            }
 
             std::apply(function, arguments);
         },
