+++ src/Coordination/Standalone/Context.cpp
@@ -458,9 +458,4 @@ const ServerSettings & Context::getServerSettings() const
     return shared->server_settings;
 }
 
-bool Context::hasTraceCollector() const
-{
-    return false;
-}
-
 }
