+++ src/Databases/MySQL/tests/gtest_mysql_binlog.cpp
@@ -1,3 +1,4 @@
+#include <Common/ThreadPool.h>
 #include <Databases/MySQL/MySQLBinlog.h>
 #include <Databases/MySQL/MySQLBinlogEventsDispatcher.h>
 #include <Databases/MySQL/MySQLBinlogClient.h>
