+++ utils/zookeeper-dump-tree/CMakeLists.txt
@@ -1,6 +1,2 @@
 clickhouse_add_executable (zookeeper-dump-tree main.cpp ${SRCS})
-target_link_libraries(zookeeper-dump-tree PRIVATE
-    clickhouse_common_zookeeper_no_log
-    clickhouse_common_io
-    dbms
-    boost::program_options)
+target_link_libraries(zookeeper-dump-tree PRIVATE clickhouse_common_zookeeper_no_log clickhouse_common_io boost::program_options)
